+++ src/main/kotlin/spring/webmvc/application/dto/command/ProductUpdateCommand.kt
@@ -6,6 +6,6 @@ open class ProductUpdateCommand(
     val category: Category,
     val name: String,
     val description: String,
-    val price: Int,
-    val quantity: Int
+    val price: Long,
+    val quantity: Long,
 )
